Extract planner info extraction logic to separate file.
